avalaibleToPlay
