+++ src/app/components/navbar/navbar.component.html
@@ -15,7 +15,7 @@
                 <app-dropdowns-notifications class="navbarItemR"></app-dropdowns-notifications>
                 <div class="navbarAvatarBtn">
                     <app-avatar class="navbarAvatar" selectedAvatar="assets\avatar\fortnite-1.jpg"
-                        userName="{{user[1]}}" profileRouting="/profile">
+                        userName="{{user.name}}" profileRouting="/profile">
                     </app-avatar>
                     <div class="btn-group dropstart">
                         <button type="button" class="btn btn-secondary dropdown-toggle" data-bs-toggle="dropdown"
@@ -23,7 +23,7 @@
                             <i class="fa-solid fa-caret-down"></i>
                         </button>
                         <ul class="dropdown-menu dropdown-menu-end">
-                            <li><a class="dropdown-item" (click)="logOut()" [routerLink]="['/landing']">Cerrar
+                            <li><a class="dropdown-item" (click)="logoutStatus()" [routerLink]="['/landing']">Cerrar
                                     sesión</a>
                             </li>
                         </ul>
